+++ modules/core/lib/src/features/features_core_presenter.dart
@@ -2,6 +2,7 @@ import 'package:dependencies/dependencies.dart';
 
 import '../models/pong_response_model.dart';
 import '../utils/erros.dart';
+import '../utils/parameters.dart';
 import '../utils/typedefs.dart';
 
 final class FeaturesCorePresenter {
@@ -45,11 +46,10 @@ final class FeaturesCorePresenter {
 
   Future<String> consumoApiRandom() async {
     final data = await _consumoApiRandomUsecase(
-      NoParams(
-        error: ApiPingError(message: "Errro ao carregar dados da API"),
+      ParametrosRandom(
+        error: ApiRandomError(message: "Errro ao carregar dados da API"),
       ),
     );
-    Logger().d(data);
     switch (data) {
       case SuccessReturn<PasswordSchemaModel>():
         return data.result.password;