+++ modules/core/lib/src/utils/erros.dart
@@ -9,3 +9,27 @@ final class ApiPingError implements AppError {
     return "ApiPingError - $message";
   }
 }
+
+final class ApiRandomError implements AppError {
+  @override
+  String message;
+  ApiRandomError({required this.message});
+  @override
+  String toString() {
+    return "ApiRandomError - $message";
+  }
+}
+
+final class ApiValidateError implements AppError {
+  final ErrorResponseModel errorResponse;
+  @override
+  String message;
+  ApiValidateError({
+    required this.message,
+    required this.errorResponse,
+  });
+  @override
+  String toString() {
+    return "ApiValidateError - $message";
+  }
+}