+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 Desenvolvimento do código fonte.
 1 - Conficuração dos and points para consumo da api.
 2 - Criação dos Models.
-3 - Criação da feature para consumo da API.
+3 - Criação da feature e testes para consumo da API endpoint pong.
+4 - Criação da feature e testes para consumo da API endpoint random.
 
 ## 0.0.1
 Configuração inicial do projeto.